+++ test_unstructured/partition/pdf_image/test_pdf.py
@@ -1517,3 +1517,52 @@ def test_document_to_element_list_sets_category_depth_titles():
     assert elements[1].metadata.category_depth == 2
     assert elements[2].metadata.category_depth is None
     assert elements[3].metadata.category_depth == 0
+
+
+@pytest.mark.parametrize("file_mode", ["filename", "rb", "spool"])
+@pytest.mark.parametrize(
+    ("strategy", "origin"),
+    # fast: can't capture the "intentionally left blank page" page
+    # others: will ignore the actual blank page
+    [
+        (PartitionStrategy.FAST,  {"pdfminer"}),
+        (PartitionStrategy.FAST,  {"pdfminer"}),
+        (PartitionStrategy.HI_RES, {"yolox", "pdfminer", "ocr_tesseract"}),
+        (PartitionStrategy.OCR_ONLY,  {"ocr_tesseract"}),
+    ],
+)
+def test_partition_pdf_with_password(
+    file_mode,
+    strategy,
+    origin,
+    filename=example_doc_path("pdf/password.pdf"),
+):
+    # Test that the partition_pdf function can handle filename
+    def _test(result):
+        # validate that the result is a non-empty list of dicts
+        assert len(result) == 1
+        assert result[0].text == 'File with password'
+
+    if file_mode == "filename":
+        result = pdf.partition_pdf(
+            filename=filename, strategy=strategy,
+            password="password"
+        )
+        _test(result)
+    elif file_mode == "rb":
+        with open(filename, "rb") as f:
+            result = pdf.partition_pdf(
+                file=f, strategy=strategy,
+                password="password"
+            )
+            _test(result)
+    else:
+        with open(filename, "rb") as test_file:
+            spooled_temp_file = SpooledTemporaryFile()
+            spooled_temp_file.write(test_file.read())
+            spooled_temp_file.seek(0)
+            result = pdf.partition_pdf(
+                file=spooled_temp_file, strategy=strategy,
+                password="password"
+            )
+            _test(result)